feat: :sparkles: Edit project
update project details

prefill all details, date and category
