+++ config/routes.rb
@@ -9,7 +9,7 @@ Rails.application.routes.draw do
   resources :sessions, only: [:new, :create]
   get 'exit', to: 'sessions#destroy', as: :logout
   resources :categories, only: :index
-  resources :projects, only: [:index, :show, :new, :create, :destroy]
+  resources :projects, only: [:index, :show, :new, :create, :destroy, :edit, :update]
   get 'myprojects', to: 'projects#myprojects', shallow: true
 
   if Rails.env.production?
